backend/requirements.txt
Django==3.2.16
djangorestframework==3.12.4
webcolors==1.11.1
psycopg2-binary==2.9.3
Pillow==9.0.0
pytest==6.2.4
pytest-django==4.4.0
PyYAML==6.0
python-dotenv
requests==2.26.0
PyJWT==2.1.0
pytest-pythonpath==0.7.3
djoser==2.2.3
django-filter==23.5
python-dotenv
django-docker-helpers